requirements.txt
biopython==1.66
numpy==1.11.0
pandas==0.18.1
python-dateutil==2.5.3
pytz==2016.4
PyYAML==3.11
six==1.10.0
